+++ @pob/eslint-config/lib/plugins/unicorn.js
@@ -2,15 +2,11 @@
 
 'use strict';
 
-const usePlugin = require('@pob/use-eslint-plugin');
-
-usePlugin(require('eslint-plugin-unicorn/package.json'));
-
 module.exports = {
   plugins: ['unicorn'],
   rules: {
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/catch-error-name.md
-    'unicorn/catch-error-name': ['error', { name: 'err' }],
+    'unicorn/catch-error-name': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/explicit-length-check.md
     'unicorn/explicit-length-check': 'error',
@@ -68,9 +64,6 @@ module.exports = {
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/error-message.md
     'unicorn/error-message': 'error',
-
-    // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/no-unsafe-regex.md
-    'unicorn/no-unsafe-regex': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-add-event-listener.md
     'unicorn/prefer-add-event-listener': 'error',
@@ -232,7 +225,7 @@ module.exports = {
     'unicorn/prefer-switch': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-node-protocol.md
-    'unicorn/prefer-node-protocol': 'off',
+    'unicorn/prefer-node-protocol': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-module.md
     // unicorn/prefer-module is enabled in node configs.
@@ -263,8 +256,7 @@ module.exports = {
     'unicorn/prefer-top-level-await': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-at.md
-    // use when node 14 support drop
-    'unicorn/prefer-at': 'off',
+    'unicorn/prefer-at': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-object-from-entries.md
     'unicorn/prefer-object-from-entries': 'error',
@@ -328,8 +320,7 @@ module.exports = {
     'unicorn/prefer-logical-operator-over-ternary': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-event-target.md
-    // TODO [engine:node@>=16]: change to error
-    'unicorn/prefer-event-target': 'warn',
+    'unicorn/prefer-event-target': 'error',
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/no-unnecessary-await.md
     'unicorn/no-unnecessary-await': 'error',
@@ -345,6 +336,9 @@ module.exports = {
 
     // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-set-size.md
     'unicorn/prefer-set-size': 'error',
+
+    // https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/docs/rules/prefer-blob-reading-methods.md
+    'unicorn/prefer-blob-reading-methods': 'error',
   },
 
   overrides: [